Handle MODIFIED events
We do this by tracking both the list of associated IPs with a host and
the hosts associated with an Ingress in a cluster. Upon receiving a
MODIFIED event, we simply delete all hosts (and their associated IPs)
from our cluster and add the current ones we see in the configuration.
Since the template is only applied after any modifications to the
dictionaries, this should lead to a consistent state.