Cluster handling: Remove startup check.
We used to list all resources on connect in order to pre-populate
everything. However, it turns out, that on a real Kubernetes cluster the
Watch will actually contain those events and this was only an effect of
the fake clientset. Remove this code and the tests for it as it actually
works in practice.