[Cluster]: Send empty update on connection failure
This changes the watch logic in the cluster handler slightly: Instead of
only sending updates when the watches have been established, we now send
empty updates in case this fails. This should make it easier to
bootstrap clusters through k8router, as it is possible to fall back to a
config where only the apiserver is reachable if ingress has not been set
up yet.